23801 - rename files and namespace
