+++ src/NHSD.GPIT.BuyingCatalogue.UI.Components/Views/Shared/Components/NhsAnchorButton/NhsAnchorButton.cshtml
@@ -1,4 +1,4 @@
-﻿@model NHSD.GPIT.BuyingCatalogue.UI.Components.Views.Shared.Components.NhsSecondaryButton.NhsAnchorButtonModel
+﻿@model NHSD.GPIT.BuyingCatalogue.UI.Components.Views.Shared.Components.NhsAnchorButton.NhsAnchorButtonModel
 
     <a class="nhsuk-button @Model.ButtonClass" href="@Model.Url" draggable="false">
         @Model.Text
